CamundaTaskConstants.java renamed to TaskAttributeConstants, annotations added to Task Types dmn
