+++ src/main/resources/dmn/wa-task-types-probate-grantofrepresentation.dmn
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<definitions xmlns="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:dmndi="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:dc="http://www.omg.org/spec/DMN/20180521/DC/" xmlns:biodi="http://bpmn.io/schema/dmn/biodi/2.0" xmlns:modeler="http://camunda.org/schema/modeler/1.0" xmlns:camunda="http://camunda.org/schema/1.0/dmn" id="Definitions_1kep5qz" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="5.34.0" modeler:executionPlatform="Camunda Cloud" modeler:executionPlatformVersion="8.0.0">
+<definitions xmlns="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:dmndi="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:dc="http://www.omg.org/spec/DMN/20180521/DC/" xmlns:biodi="http://bpmn.io/schema/dmn/biodi/2.0" xmlns:modeler="http://camunda.org/schema/modeler/1.0" xmlns:camunda="http://camunda.org/schema/1.0/dmn" id="Definitions_1kep5qz" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="5.33.1" modeler:executionPlatform="Camunda Cloud" modeler:executionPlatformVersion="8.0.0">
   <decision id="wa-task-types-probate-grantofrepresentation" name="Task Types DMN" camunda:historyTimeToLive="P90D">
     <decisionTable id="DecisionTable_17knzal" hitPolicy="COLLECT">
       <input id="Input_1">
@@ -10,6 +10,7 @@
       <output id="Output_1" name="taskTypeId" typeRef="string" biodi:width="359" />
       <output id="OutputClause_0aeuwfu" name="taskTypeName" typeRef="string" biodi:width="489" />
       <rule id="DecisionRule_1s0e55i">
+        <description>EDCP</description>
         <inputEntry id="UnaryTests_159d5fs">
           <text></text>
         </inputEntry>
@@ -21,6 +22,7 @@
         </outputEntry>
       </rule>
       <rule id="DecisionRule_1gmuhtt">
+        <description>EDCAW</description>
         <inputEntry id="UnaryTests_0vbok3e">
           <text></text>
         </inputEntry>
@@ -32,6 +34,7 @@
         </outputEntry>
       </rule>
       <rule id="DecisionRule_1oeofpy">
+        <description>EDCI</description>
         <inputEntry id="UnaryTests_1pyk95b">
           <text></text>
         </inputEntry>
